use state machine to change man's action

+++ GameLib/GameLib.h
@@ -9,9 +9,8 @@ int Game_Shutdown(void * parms = NULL, int num_parms = 0);
 int Game_Main(void * parms = NULL, int num_parms = 0);
 int Game_Debug_Textout();
 
-
-extern	unsigned int		Mouse_X;
-extern	unsigned int		Mouse_Y;
+extern	long				Mouse_X;
+extern	long				Mouse_Y;
 extern	bool				bLBUTTONDOWN;
 extern	HWND				main_window_handle; // save the window handle
